chore: IntoJArray is part of owned
