+++ src/arrays/mod.rs
@@ -16,7 +16,7 @@ use thiserror::Error;
 
 pub use arrayable::Arrayable;
 pub use cow::{CowArrayD, JArrayCow};
-pub use owned::JArray;
+pub use owned::{IntoJArray, JArray};
 
 // TODO: https://code.jsoftware.com/wiki/Vocabulary/ErrorMessages
 #[derive(Debug, Error)]
@@ -251,38 +251,6 @@ impl_empty!(BigRational, BigRational::zero());
 impl_empty!(f64, 0.);
 impl_empty!(Complex64, Complex64::zero());
 impl_empty!(Word, Noun(BoolArray(Array::from_elem(IxDyn(&[0]), 0))));
-
-pub trait IntoJArray {
-    fn into_jarray(self) -> JArray;
-    fn into_noun(self) -> Word
-    where
-        Self: Sized,
-    {
-        Word::Noun(self.into_jarray())
-    }
-}
-
-macro_rules! impl_into_jarray {
-    ($t:ty, $j:path) => {
-        impl IntoJArray for $t {
-            /// free for ArrayD<>, clones for unowned CowArrayD<>
-            fn into_jarray(self) -> JArray {
-                $j(self.into_owned())
-            }
-        }
-    };
-}
-
-// these also cover the CowArrayD<> conversions because both are just aliases
-// for ArrayBase<T> and the compiler lets us get away without lifetimes for some reason.
-impl_into_jarray!(ArrayD<u8>, JArray::BoolArray);
-impl_into_jarray!(ArrayD<char>, JArray::CharArray);
-impl_into_jarray!(ArrayD<i64>, JArray::IntArray);
-impl_into_jarray!(ArrayD<BigInt>, JArray::ExtIntArray);
-impl_into_jarray!(ArrayD<BigRational>, JArray::RationalArray);
-impl_into_jarray!(ArrayD<f64>, JArray::FloatArray);
-impl_into_jarray!(ArrayD<Complex64>, JArray::ComplexArray);
-impl_into_jarray!(ArrayD<Word>, JArray::BoxArray);
 
 impl Word {
     pub fn noun<T>(v: impl Arrayable<T>) -> Result<Word>
